feat 🦄: add CREDIT/DEBIT rule types
Added CREDIT or DEBIT type to rules determining the sign of the
contribution of each transaction in the rule.
CREDIT rules have same sign as transactions.
DEBIT rules have swapped signs.
CREDIT rules are for example the total by store.
DEBIT rules are for example the total in VATs of a store.